+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Features
 
-- ASAB WebUI submodule version bump [[80151df](https://github.com/TeskaLabs/asab-webui/commit/80151df2eb26bc683a6b5d31bfed9a0acd879c35)] commit (INDIGO Sprint 230526, [!37](https://github.com/TeskaLabs/seacat-admin-webui/pull/37))
+- ASAB WebUI submodule version bump [[5f22a1c](https://github.com/TeskaLabs/asab-webui/commit/5f22a1cf595562032759852dd5233af892e596fa)] commit (INDIGO Sprint 230526, [!21](https://github.com/TeskaLabs/seacat-admin-webui/pull/21))
 
 - Changing logo based on app's theme (INDIGO Sprint 230428, [!12](https://github.com/TeskaLabs/seacat-admin-webui/pull/12))
 
 - Pagination via `More..` button in dropdowns. Clients refactored to easily maintainable css grid. (INDIGO 230526, [!35](https://github.com/TeskaLabs/seacat-admin-webui/pull/35))
+
+- Bulk assignment sceen for miltuple tenants & roles to credentials assignment. (INDIGO 230526, [!21](https://github.com/TeskaLabs/seacat-admin-webui/pull/21))
 
 ### Refactoring